# openapi.yaml
openapi: 3.0.3
info:
  title: Admin Dashboard API
  description: API cho hệ thống đặt hàng đa NCC (Multi-Supplier Order). Chi tiết xem API-FE-LIEN-KET.md
  version: 1.0.0
servers:
  - url: /api
    description: Base path API
...
paths:
...
  /orders:
    get:
      tags: [Orders]
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: storeId
          in: query
          schema: { type: integer }
        - name: supplierId
          in: query
          schema: { type: integer }
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Order' }
    post:
      tags: [Orders]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderCreateInput' }
      security: [{ Bearer: [] }]
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
  /orders/{id}:
    get:
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
  /orders/{id}/accept:
    patch:
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
  /orders/{id}/reject:
    patch:
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
  /orders/{id}/confirm-total:
    patch:
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token trả về từ POST /auth/login
  schemas:
...
    OrderItem:
      type: object
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unit: { type: string }
        price: { type: number }
        discountAmount: { type: number }
        taxPercent: { type: number }
        taxAmount: { type: number }
    OrderItemInput:
      type: object
      properties:
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unit: { type: string }
        price: { type: number }
        discountAmount: { type: number, default: 0 }
        taxPercent: { type: number, default: 0 }
        taxAmount: { type: number, default: 0 }
    ReceiveImage:
      type: object
      properties:
        id: { type: integer }
        type: { type: string, enum: [received, invoice] }
        imageUrl: { type: string }
        fileName: { type: string }
    OrderSupplier:
      type: object
      properties:
        id: { type: integer }
        orderId: { type: integer }
        supplierId: { type: integer }
        supplierName: { type: string }
        status: { type: string }
        expectedDeliveryDate: { type: string, format: date }
        actualDeliveryDate: { type: string, format: date, nullable: true }
        confirmDate: { type: string, format: date, nullable: true }
        note: { type: string }
        totalAmount: { type: number }
        taxAmount: { type: number }
        paymentStatus: { type: string }
        paidAmount: { type: number }
        paidDate: { type: string, format: date, nullable: true }
        receivedBy: { type: integer, nullable: true }
        receivedDate: { type: string, format: date-time, nullable: true }
        orderItems:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
        receiveImages:
          type: array
          items: { $ref: '#/components/schemas/ReceiveImage' }
    OrderSupplierCreateInput:
      type: object
      properties:
        supplierId: { type: integer }
        orderItems:
          type: array
          items: { $ref: '#/components/schemas/OrderItemInput' }
    Order:
      type: object
      properties:
        id: { type: integer }
        storeId: { type: integer }
        storeName: { type: string }
        status: { type: string }
        orderDate: { type: string, format: date }
        expectedDeliveryDate: { type: string, format: date, nullable: true }
        totalItemCount: { type: integer }
        totalAmount: { type: number }
        createdBy: { type: integer }
        createdDate: { type: string, format: date-time }
        orderSuppliers:
          type: array
          items: { $ref: '#/components/schemas/OrderSupplier' }
    OrderCreateInput:
      type: object
      required: [storeId, orderSuppliers]
      properties:
        storeId: { type: integer }
        expectedDeliveryDate: { type: string, format: date, nullable: true }
        orderSuppliers:
          type: array
          items: { $ref: '#/components/schemas/OrderSupplierCreateInput' }